+++ server/src/com/openrsc/server/database/impl/mysql/MySqlGameDatabase.java
@@ -37,7 +37,7 @@ public class MySqlGameDatabase extends JDBCDatabase {
 	private final QueriesManager queriesManager;
 	private final Queries queries;
 	private final MySqlQueries mySqlQueries;
-	private final Set<Integer> itemIDList;
+	private final Set<Long> itemIDList;
 	private final JDBCDatabaseConnection connection;
 
 	public MySqlGameDatabase(final Server server) {
@@ -318,18 +318,18 @@ public class MySqlGameDatabase extends JDBCDatabase {
 	}
 
 	@Override
-	public Integer[] queryInUseItemIds() throws GameDatabaseException {
-		final ArrayList<Integer> inUseItemIds = new ArrayList<>();
+	public Long[] queryInUseItemIds() throws GameDatabaseException {
+		final ArrayList<Long> inUseItemIds = new ArrayList<>();
 		try (final PreparedStatement statement = getConnection().prepareStatement(getMySqlQueries().inUseItemIds);
 			 final ResultSet result = statement.executeQuery()) {
 
 			while (result.next()) {
-				inUseItemIds.add(result.getInt("itemID"));
+				inUseItemIds.add(result.getLong("itemID"));
 			}
 		} catch (final SQLException ex) {
 			throw new GameDatabaseException(MySqlGameDatabase.class, ex.getMessage());
 		}
-		return inUseItemIds.toArray(new Integer[0]);
+		return inUseItemIds.toArray(new Long[0]);
 	}
 
 	@Override
@@ -1607,11 +1607,11 @@ public class MySqlGameDatabase extends JDBCDatabase {
 			updateLongs(getMySqlQueries().save_DeleteInv, playerId);
 			for (final PlayerInventory item : inventory) {
 				statement.setInt(1, playerId);
-				statement.setInt(2, item.itemId);
+				statement.setLong(2, item.itemId);
 				statement.setInt(3, item.slot);
 				statement.addBatch();
 
-				statement2.setInt(1, item.itemId);
+				statement2.setLong(1, item.itemId);
 				statement2.setInt(2, item.catalogID);
 				statement2.setInt(3, item.amount);
 				statement2.setInt(4, item.noted ? 1 : 0);
@@ -1636,10 +1636,10 @@ public class MySqlGameDatabase extends JDBCDatabase {
 			updateLongs(getMySqlQueries().save_DeleteEquip, playerId);
 			for (final PlayerEquipped item : equipment) {
 				statement.setInt(1, playerId);
-				statement.setInt(2, item.itemId);
+				statement.setLong(2, item.itemId);
 				statement.addBatch();
 
-				statement2.setInt(1, item.itemId);
+				statement2.setLong(1, item.itemId);
 				statement2.setInt(2, item.itemStatus.getCatalogId());
 				statement2.setInt(3, item.itemStatus.getAmount());
 				statement2.setInt(4, item.itemStatus.getNoted() ? 1 : 0);
@@ -1666,11 +1666,11 @@ public class MySqlGameDatabase extends JDBCDatabase {
 				int slot = 0;
 				for (final PlayerBank item : bank) {
 					statement.setInt(1, playerId);
-					statement.setInt(2, item.itemId);
+					statement.setLong(2, item.itemId);
 					statement.setInt(3, slot++);
 					statement.addBatch();
 
-					statement2.setInt(1, item.itemId);
+					statement2.setLong(1, item.itemId);
 					statement2.setInt(2, item.itemStatus.getCatalogId());
 					statement2.setInt(3, item.itemStatus.getAmount());
 					statement2.setInt(4, item.itemStatus.getNoted() ? 1 : 0);
@@ -1914,13 +1914,13 @@ public class MySqlGameDatabase extends JDBCDatabase {
 	}
 
 	@Override
-	public int queryMaxItemID() throws GameDatabaseException {
-		int maxId = 0;
+	public long queryMaxItemID() throws GameDatabaseException {
+		long maxId = 0;
 		try (final PreparedStatement statement = getConnection().prepareStatement(getMySqlQueries().max_itemStatus);
 			 final ResultSet result = statement.executeQuery()) {
 
 			if (result.next()) {
-				maxId = result.getInt("itemID");
+				maxId = result.getLong("itemID");
 			}
 		} catch (final SQLException ex) {
 			// Convert SQLException to a general usage exception
@@ -1957,7 +1957,7 @@ public class MySqlGameDatabase extends JDBCDatabase {
 		try (final PreparedStatement statement = getConnection().prepareStatement(getMySqlQueries().save_ItemPurge)) {
 			purgeItemID(item.getItemId());
 
-			statement.setInt(1, item.getItemId());
+			statement.setLong(1, item.getItemId());
 			statement.executeUpdate();
 		} catch (final SQLException ex) {
 			// Convert SQLException to a general usage exception
@@ -1975,7 +1975,7 @@ public class MySqlGameDatabase extends JDBCDatabase {
 				statement.setInt(2, item.getNoted() ? 1 : 0);
 				statement.setInt(3, item.isWielded() ? 1 : 0);
 				statement.setInt(4, item.getItemStatus().getDurability());
-				statement.setInt(5, item.getItemId());
+				statement.setLong(5, item.getItemId());
 
 				statement.executeUpdate();
 			}
@@ -1989,13 +1989,13 @@ public class MySqlGameDatabase extends JDBCDatabase {
 	public void queryInventoryAdd(final int playerId, final Item item, final int slot) throws GameDatabaseException {
 		synchronized (itemIDList) {
 			try (final PreparedStatement statement = getConnection().prepareStatement(getMySqlQueries().save_InventoryAdd)) {
-				int itemId = item.getItemId();
+				long itemId = item.getItemId();
 				if (itemId == Item.ITEM_ID_UNASSIGNED) {
 					itemId = assignItemID(item);
 				}
 
 				statement.setInt(1, playerId);
-				statement.setInt(2, itemId);
+				statement.setLong(2, itemId);
 				statement.setInt(3, slot);
 				statement.executeUpdate();
 			} catch (final SQLException ex) {
@@ -2010,7 +2010,7 @@ public class MySqlGameDatabase extends JDBCDatabase {
 		synchronized (itemIDList) {
 			try (final PreparedStatement statement = getConnection().prepareStatement(getMySqlQueries().save_InventoryRemove)) {
 				statement.setInt(1, playerId);
-				statement.setInt(2, item.getItemId());
+				statement.setLong(2, item.getItemId());
 				statement.executeUpdate();
 			} catch (final SQLException ex) {
 				// Convert SQLException to a general usage exception
@@ -2023,13 +2023,13 @@ public class MySqlGameDatabase extends JDBCDatabase {
 	public void queryEquipmentAdd(final int playerId, final Item item) throws GameDatabaseException {
 		synchronized (itemIDList) {
 			try (final PreparedStatement statement = getConnection().prepareStatement(getMySqlQueries().save_EquipmentAdd)) {
-				int itemId = item.getItemId();
+				long itemId = item.getItemId();
 				if (itemId == Item.ITEM_ID_UNASSIGNED) {
 					itemId = assignItemID(item);
 				}
 
 				statement.setInt(1, playerId);
-				statement.setInt(2, itemId);
+				statement.setLong(2, itemId);
 				statement.executeUpdate();
 			} catch (final SQLException ex) {
 				// Convert SQLException to a general usage exception
@@ -2043,7 +2043,7 @@ public class MySqlGameDatabase extends JDBCDatabase {
 		synchronized (itemIDList) {
 			try (final PreparedStatement statement = getConnection().prepareStatement(getMySqlQueries().save_EquipmentRemove)) {
 				statement.setInt(1, playerId);
-				statement.setInt(2, item.getItemId());
+				statement.setLong(2, item.getItemId());
 				statement.executeUpdate();
 			} catch (final SQLException ex) {
 				// Convert SQLException to a general usage exception
@@ -2056,13 +2056,13 @@ public class MySqlGameDatabase extends JDBCDatabase {
 	public void queryBankAdd(final int playerId, final Item item, final int slot) throws GameDatabaseException {
 		synchronized (itemIDList) {
 			try (final PreparedStatement statement = getConnection().prepareStatement(getMySqlQueries().save_BankAdd)) {
-				int itemId = item.getItemId();
+				long itemId = item.getItemId();
 				if (itemId == Item.ITEM_ID_UNASSIGNED) {
 					itemId = assignItemID(item);
 				}
 
 				statement.setInt(1, playerId);
-				statement.setInt(2, itemId);
+				statement.setLong(2, itemId);
 				statement.setInt(3, slot);
 				statement.executeUpdate();
 			} catch (final SQLException ex) {
@@ -2077,7 +2077,7 @@ public class MySqlGameDatabase extends JDBCDatabase {
 		synchronized (itemIDList) {
 			try (final PreparedStatement statement = getConnection().prepareStatement(getMySqlQueries().save_BankRemove)) {
 				statement.setInt(1, playerId);
-				statement.setInt(2, item.getItemId());
+				statement.setLong(2, item.getItemId());
 				statement.executeUpdate();
 			} catch (final SQLException ex) {
 				// Convert SQLException to a general usage exception
@@ -2095,7 +2095,7 @@ public class MySqlGameDatabase extends JDBCDatabase {
 				}
 				statement.setInt(1, item.getAmount() - amountToRemove);
 				statement.setInt(2, playerId);
-				statement.setInt(3, item.getItemId());
+				statement.setLong(3, item.getItemId());
 				statement.executeUpdate();
 			} catch (final SQLException ex) {
 				// Convert SQLException to a general usage exception
@@ -2532,13 +2532,13 @@ public class MySqlGameDatabase extends JDBCDatabase {
 		return mySqlQueries;
 	}
 
-	public Set<Integer> getItemIDList() {
+	public Set<Long> getItemIDList() {
 		return this.itemIDList;
 	}
 
-	public int addItemToPlayer(final Item item) {
+	public long addItemToPlayer(final Item item) {
 		try {
-			int itemId = item.getItemId();
+			long itemId = item.getItemId();
 			if (itemId == Item.ITEM_ID_UNASSIGNED) {
 				return assignItemID(item);
 			}
@@ -2642,20 +2642,20 @@ public class MySqlGameDatabase extends JDBCDatabase {
 		}
 	}
 
-	public int assignItemID(final Item item) throws GameDatabaseException {
+	public long assignItemID(final Item item) throws GameDatabaseException {
 		synchronized (itemIDList) {
-			int itemId = itemCreate(item);
+			long itemId = itemCreate(item);
 			item.setItemId(itemId);
 			itemIDList.add(itemId);
 			return itemId;
 		}
 	}
 
-	protected void purgeItemID(final int itemID) {
+	protected void purgeItemID(final long itemID) {
 		synchronized (itemIDList) {
-			Iterator<Integer> iterator = itemIDList.iterator();
+			Iterator<Long> iterator = itemIDList.iterator();
 			while (iterator.hasNext()) {
-				Integer listID = iterator.next();
+				Long listID = iterator.next();
 				if (listID == itemID) {
 					iterator.remove();
 					return;
