+++ server/src/com/openrsc/server/database/struct/PlayerInventory.java
@@ -3,7 +3,7 @@ package com.openrsc.server.database.struct;
 import com.openrsc.server.model.container.Item;
 
 public class PlayerInventory {
-	public int itemId;
+	public long itemId;
 	public boolean wielded;
 	public int slot;
 	public Item item;
